+++ lib/api/core/plugins/pluginsManager.js
@@ -22,7 +22,7 @@
 'use strict';
 
 const
-  debug = require('../../../debug')('kuzzle:plugins'),
+  debug = require('../../../kuzzleDebug')('kuzzle:plugins'),
   PluginContext = require('./pluginContext'),
   PrivilegedPluginContext = require('./privilegedPluginContext'),
   async = require('async'),
