+++ lib/services/broker/index.js
@@ -20,7 +20,7 @@
  */
 
 const
-  debug = require('../../debug')('kuzzle:broker'),
+  debug = require('../../kuzzleDebug')('kuzzle:broker'),
   WSBrokerClient = require('./wsBrokerClient'),
   WSBrokerServer = require('./wsBrokerServer');
