+++ lib/services/broker/wsBrokerClient.js
@@ -22,7 +22,7 @@
 'use strict';
 
 const
-  debug = require('../../debug')('kuzzle:broker:client'),
+  debug = require('../../kuzzleDebug')('kuzzle:broker:client'),
   path = require('path'),
   Bluebird = require('bluebird'),
   KuzzleInternalError = require('kuzzle-common-objects').errors.InternalError,
