+++ lib/services/broker/wsBrokerServer.js
@@ -22,7 +22,7 @@
 'use strict';
 
 const
-  debug = require('../../debug')('kuzzle:broker:server'),
+  debug = require('../../kuzzleDebug')('kuzzle:broker:server'),
   path = require('path'),
   Bluebird = require('bluebird'),
   CircularList = require('easy-circular-list'),
