+++ lib/services/elasticsearch.js
@@ -22,7 +22,7 @@
 'use strict';
 
 const
-  debug = require('../debug')('kuzzle:services:elasticsearch'),
+  debug = require('../kuzzleDebug')('kuzzle:services:elasticsearch'),
   _ = require('lodash'),
   Bluebird = require('bluebird'),
   Service = require('./service'),
